% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset.R
\name{inset}
\alias{inset}
\alias{inset_position}
\alias{inset_background}
\alias{inset_templates}
\title{Meme plot inset templates}
\usage{
inset_position(type = "default")

inset_background(type = "default")

inset_templates(type)
}
\arguments{
\item{type}{character, name of template.}
}
\value{
a list of arguments passed to either \code{inset_pos} or \code{inset_background} in \code{meme}.
}
\description{
Templates for the position and background of a ggplot inset.
}
\details{
\code{inset_position} and \code{inset_background} assist with some basic options for position and background of the optional ggplot inset graphic.
\code{inset_templates} can be used to view the available templates. See examples.
If a template is not available to suit your needs, provide your own argument list to \code{meme}.
}
\examples{
inset_templates("position")
inset_templates("background")
inset_position()
inset_background()
}
