% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memery.R
\docType{package}
\name{memery}
\alias{memery}
\alias{memery-package}
\title{memery: Internet memes for data analysts.}
\description{
The memery package is used for generating internet memes that may optionally include a superimposed inset plot.
It also offers the classic Impact font family with shadow/outline effect for the meme text label.
}
