% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{memetext_position}
\alias{memetext_position}
\title{Default meme text position}
\usage{
memetext_position(n)
}
\arguments{
\item{n}{integer, number of meme text labels.}
}
\value{
a list of meme text label position arguments: width, height, and x and y coordinates.
}
\description{
Convenience function for meme text position in a meme plot.
}
\details{
This function takes an integer 1 or 2 and returns, respectively, top or symmetrical top and bottom meme text position arguments.
This function is used as the default for the \code{label_pos} argument in \code{meme}.
It is provided if you do not want to bother with specifying coordinates and exact placement does not matter.
It is not intended for placement of more than two meme text labels and any value other than 1 or 2 returns an error.
}
