% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset.R
\name{memetheme}
\alias{memetheme}
\title{Default meme theme}
\usage{
memetheme(base_size = 14, base_family = "", base_col = "white")
}
\arguments{
\item{base_size}{numeric, the base size.}

\item{base_family}{character, the base font family.}

\item{base_col}{character, the base color for all title text and axis lines and ticks.}
}
\value{
a ggplot2 theme.
}
\description{
The default ggplot2 theme for meme plots.
}
